little cleaning and refactoring
