+++ src/main/java/net/imglib2/realtransform/RealTransformSequenceAdapter.java
@@ -30,7 +30,7 @@ package net.imglib2.realtransform;
 
 import com.google.gson.*;
 import org.scijava.plugin.Plugin;
-import sc.fiji.serializers.IClassRuntimeAdapter;
+import sc.fiji.persist.IClassRuntimeAdapter;
 
 import java.lang.reflect.Type;
 
@@ -70,17 +70,8 @@ public class RealTransformSequenceAdapter implements IClassRuntimeAdapter<RealTr
         RealTransformSequence rts = new RealTransformSequence();
 
         for (int iTransform = 0; iTransform<nTransform; iTransform++) {
-            // Special case in order to deserialize directly
-            // affine transforms to AffineTransform3D objects
-            // JsonObject jsonObj = obj.get("realTransform_"+iTransform).getAsJsonObject();
-            /*if (jsonObj.has("affinetransform3d")) {
-                AffineTransform3D at3D = jsonDeserializationContext.deserialize(obj.get("realTransform_"+iTransform), AffineTransform3D.class);
-                rts.add(at3D);
-            } else*/
-            //{
             RealTransform transform = jsonDeserializationContext.deserialize(obj.get("realTransform_"+iTransform), RealTransform.class);
             rts.add(transform);
-            //}
         }
 
         return rts;
@@ -90,8 +81,6 @@ public class RealTransformSequenceAdapter implements IClassRuntimeAdapter<RealTr
     public JsonElement serialize(RealTransformSequence rts, Type type, JsonSerializationContext jsonSerializationContext) {
 
         JsonObject obj = new JsonObject();
-
-        //obj.addProperty("type", RealTransformSequence.class.getSimpleName());
 
         obj.addProperty("size", rts.transforms.size());
