+++ src/main/java/sc/fiji/bdvpg/scijava/services/SourceAndConverterBdvDisplayService.java
@@ -60,7 +60,7 @@ import java.util.stream.Collectors;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
-import sc.fiji.serializers.ScijavaGsonHelper;
+import sc.fiji.persist.ScijavaGsonHelper;
 
 import javax.swing.tree.DefaultTreeModel;
