+++ src/main/java/sc/fiji/persist/DefaultScijavaAdapterService.java
@@ -1,4 +1,4 @@
-package sc.fiji.serializers;
+package sc.fiji.persist;
 
 import org.scijava.Context;
 import org.scijava.Priority;
