+++ src/main/java/sc/fiji/persist/IClassAdapter.java
@@ -1,4 +1,4 @@
-package sc.fiji.serializers;
+package sc.fiji.persist;
 
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonSerializer;
