+++ src/main/java/sc/fiji/persist/IObjectScijavaAdapter.java
@@ -1,4 +1,4 @@
-package sc.fiji.serializers;
+package sc.fiji.persist;
 
 import org.scijava.plugin.SciJavaPlugin;
