+++ src/main/java/sc/fiji/persist/IObjectScijavaAdapterService.java
@@ -1,4 +1,4 @@
-package sc.fiji.serializers;
+package sc.fiji.persist;
 
 import org.scijava.plugin.PTService;
 import org.scijava.plugin.PluginInfo;
